requirements.txt
requests==2.25.1
2captcha-python==1.1.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dotenv==0.17.1
selenium==3.141.0
urllib3==1.26.4
